pyproject.toml
[project]
name = "gsf-server"
version = "0.1.0"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "nemo-retriever",
    "neo4j>=6.1.0",
    "pandas>=2.0,<3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.32.0",
]

[tool.uv]
package = false

[tool.uv.sources]
nemo-retriever = { git = "https://github.com/NVIDIA/NeMo-Retriever.git", subdirectory = "nemo_retriever" }

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "ruff>=0.15.9",
]
